# Conflicts:
#	Resources/Prototypes/Entities/Objects/Consumable/Food/ingredients.yml
#	Resources/Prototypes/tags.yml

+++ Content.Server/Botany/Systems/BotanySystem.Produce.cs
@@ -18,7 +18,7 @@ public sealed partial class BotanySystem
             var amount = FixedPoint2.New(quantity.Min);
             if (quantity.PotencyDivisor > 0 && seed.Potency > 0)
                 amount += FixedPoint2.New(seed.Potency / quantity.PotencyDivisor);
-            amount = FixedPoint2.New((int) MathHelper.Clamp(amount.Float(), quantity.Min, quantity.Max));
+            amount = FixedPoint2.New(MathHelper.Clamp(amount.Float(), quantity.Min, quantity.Max));
             solutionContainer.MaxVolume += amount;
             solutionContainer.AddReagent(chem, amount);
         }
